Collapse `eventstore` into `persistence`.
