+++ persistence/operation.go
@@ -4,27 +4,8 @@ import (
 	"context"
 	"strings"
 
-	"github.com/dogmatiq/infix/draftspecs/envelopespec"
 	"github.com/dogmatiq/infix/persistence/subsystem/queuestore"
 )
-
-// Operation is a persistence operation that can be performed as part of an
-// atomic batch.
-type Operation interface {
-	// AcceptVisitor calls the appropriate visit method on the given visitor.
-	AcceptVisitor(context.Context, OperationVisitor) error
-
-	// entityKey a value that identifies the persisted "entity" that the
-	// operation manipulates. No two operations in the same batch may operate
-	// upon the same entity.
-	entityKey() entityKey
-}
-
-// SaveEvent is a persistence operation that persists an event message.
-type SaveEvent struct {
-	// Envelope is the envelope containing the event to persist.
-	Envelope *envelopespec.Envelope
-}
 
 // SaveQueueItem is a persistence operation that creates or updates an item on
 // the message queue.
@@ -48,25 +29,6 @@ type RemoveQueueItem struct {
 	Item queuestore.Item
 }
 
-// OperationVisitor visits persistence operations.
-type OperationVisitor interface {
-	VisitSaveAggregateMetaData(context.Context, SaveAggregateMetaData) error
-	VisitSaveEvent(context.Context, SaveEvent) error
-	VisitSaveQueueItem(context.Context, SaveQueueItem) error
-	VisitRemoveQueueItem(context.Context, RemoveQueueItem) error
-	VisitSaveOffset(context.Context, SaveOffset) error
-}
-
-// AcceptVisitor calls v.VisitSaveAggregateMetaData().
-func (op SaveAggregateMetaData) AcceptVisitor(ctx context.Context, v OperationVisitor) error {
-	return v.VisitSaveAggregateMetaData(ctx, op)
-}
-
-// AcceptVisitor calls v.VisitSaveEvent().
-func (op SaveEvent) AcceptVisitor(ctx context.Context, v OperationVisitor) error {
-	return v.VisitSaveEvent(ctx, op)
-}
-
 // AcceptVisitor calls v.VisitSaveQueueItem().
 func (op SaveQueueItem) AcceptVisitor(ctx context.Context, v OperationVisitor) error {
 	return v.VisitSaveQueueItem(ctx, op)
@@ -75,11 +37,6 @@ func (op SaveQueueItem) AcceptVisitor(ctx context.Context, v OperationVisitor) e
 // AcceptVisitor calls v.VisitRemoveQueueItem().
 func (op RemoveQueueItem) AcceptVisitor(ctx context.Context, v OperationVisitor) error {
 	return v.VisitRemoveQueueItem(ctx, op)
-}
-
-// AcceptVisitor calls v.VisitSaveOffset().
-func (op SaveOffset) AcceptVisitor(ctx context.Context, v OperationVisitor) error {
-	return v.VisitSaveOffset(ctx, op)
 }
 
 // entityKey uniquely identifies the entity that is affected by an operation.
@@ -91,22 +48,10 @@ func (k entityKey) String() string {
 	)
 }
 
-func (op SaveAggregateMetaData) entityKey() entityKey {
-	return entityKey{"aggregate", op.MetaData.HandlerKey, op.MetaData.InstanceID}
-}
-
-func (op SaveEvent) entityKey() entityKey {
-	return entityKey{"event", op.Envelope.MetaData.MessageId}
-}
-
 func (op SaveQueueItem) entityKey() entityKey {
 	return entityKey{"queue", op.Item.ID()}
 }
 
 func (op RemoveQueueItem) entityKey() entityKey {
 	return entityKey{"queue", op.Item.ID()}
-}
-
-func (op SaveOffset) entityKey() entityKey {
-	return entityKey{"offset", op.ApplicationKey}
 }
