+++ persistence/subsystem/eventstore/item.go
@@ -1,16 +1,5 @@
 package eventstore
 
-import (
-	"github.com/dogmatiq/infix/draftspecs/envelopespec"
-)
+import "github.com/dogmatiq/infix/persistence"
 
-// Item is a message persisted in the event store.
-type Item struct {
-	Offset   uint64
-	Envelope *envelopespec.Envelope
-}
-
-// ID returns the ID of the message.
-func (i *Item) ID() string {
-	return i.Envelope.MetaData.MessageId
-}
+type Item = persistence.Event
